requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

python-docx==1.1.0
openpyxl==3.1.2
Pillow==10.1.0
pytesseract==0.3.10

gtts==2.4.0
pyttsx3==2.90

google-cloud-storage==2.14.0
requests==2.31.0

redis==5.0.1
celery==5.3.4

python-json-logger==2.0.7

pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
faker==20.1.0
aiosqlite==0.19.0

black==23.11.0
flake8==6.1.0
mypy==1.7.1

fsrs-optimizer==4.8.0
